+++ vector3.h
@@ -6,18 +6,18 @@
 
 class Vector3 {
 public:
-    float e[3];
+    double e[3];
 
     Vector3() : e{ 0, 0, 0 } {}
-    Vector3(float e0, float e1, float e2) : e{ e0, e1, e2 } {}
+    Vector3(double e0, double e1, double e2) : e{ e0, e1, e2 } {}
 
-    float x() const { return e[0]; }
-    float y() const { return e[1]; }
-    float z() const { return e[2]; }
+    double x() const { return e[0]; }
+    double y() const { return e[1]; }
+    double z() const { return e[2]; }
 
     Vector3 operator-() const { return Vector3(-e[0], -e[1], -e[2]); }
-    float operator[](int i) const { return e[i]; } /* Unmodifiable value */
-    float& operator[](int i) { return e[i]; }      /* Modifiable reference */
+    double operator[](int i) const { return e[i]; } /* Unmodifiable value */
+    double& operator[](int i) { return e[i]; }      /* Modifiable reference */
 
     Vector3& operator+=(const Vector3& other) {
         e[0] += other[0];
@@ -26,17 +26,17 @@ public:
         return *this;
     }
 
-    Vector3& operator*=(const float f) {
+    Vector3& operator*=(const double f) {
         e[0] *= f;
         e[1] *= f;
         e[2] *= f;
         return *this;
     }
 
-    Vector3& operator/=(const float f) { return *this *= 1/f; }
+    Vector3& operator/=(const double f) { return *this *= 1/f; }
 
-    float length() const { return std::sqrt(length_squared()); }
-    float length_squared() const { return e[0] * e[0] + e[1] * e[1] + e[2] * e[2]; }
+    double length() const { return std::sqrt(length_squared()); }
+    double length_squared() const { return e[0] * e[0] + e[1] * e[1] + e[2] * e[2]; }
 };
 
 using Point3 = Vector3;
@@ -62,20 +62,20 @@ inline Vector3 operator-(const Vector3 &x, const Vector3 &y) {
 inline Vector3 operator*(const Vector3 &x, const Vector3 &y) {
     return Vector3(x.e[0] * y.e[0], x.e[1] * y.e[1], x.e[2] * y.e[2]);
 }
-inline Vector3 operator*(const float x, const Vector3 &y) {
+inline Vector3 operator*(const double x, const Vector3 &y) {
     return Vector3(x * y.e[0], x * y.e[1], x * y.e[2]);
 }
-inline Vector3 operator*(const Vector3 &x, const float y) {
+inline Vector3 operator*(const Vector3 &x, const double y) {
     return y * x;
 }
 
 // Divide
-inline Vector3 operator/(const Vector3 &x, const float y) {
+inline Vector3 operator/(const Vector3 &x, const double y) {
     return (1/y) * x;
 }
 
 // Dot product
-inline float dot(const Vector3 &x, const Vector3 &y) {
+inline double dot(const Vector3 &x, const Vector3 &y) {
     return x.e[0] * y.e[0] + x.e[1] * y.e[1] + x.e[2] * y.e[2];
 }
 // Cross product
